Fix: PtzStatusBodyDto Pan/Tilt/Zoom 타입 double → int (설계 문서 §8.3.1 정합)
설계 문서 기준 PTZ 값은 정수형 카메라 내부 단위이므로 double → int 수정.
A2.17-1 테스트를 int 값(1000, 5000, 2000)으로 업데이트 및 타입 검증 추가.
172/172 Green.

+++ Ironwall.Dotnet.Libraries.Messages/Dto/Brokers/PtzStatusBodyDto.cs
@@ -8,11 +8,11 @@ public class PtzStatusBodyDto
     public int CameraId { get; set; }
 
     [JsonProperty("pan")]
-    public double Pan { get; set; }
+    public int Pan { get; set; }
 
     [JsonProperty("tilt")]
-    public double Tilt { get; set; }
+    public int Tilt { get; set; }
 
     [JsonProperty("zoom")]
-    public double Zoom { get; set; }
+    public int Zoom { get; set; }
 }
